Reproducible default thumbnail/splash selection based on the hash of the post title.
Rationale:

Random selection has the drawback that singe and list views will show
different thumbnails. It makes sense to select the thumbnail by markers
rooted in the post itself.

Of the two obvious candidates, title and
content, the title is less likely to change over the lifetime of a post
than the content. We don't want content edits to result in a new
post thumbnail. Hence we select the title as a marker.

How to compute the thumb index from the title?
First, we compute an md5 hash, so we make sure even slight title diffs
lead to significant hash code diffs. Then, we convert that hash to int.
We can't do that directly though, because the md5 code (a hex string) is
does not fit into a go int. Hence we only take the last 7 (LSB-) digits.
We convert them to an int and modulo them by the count of default thumbgs